pyproject.toml
[project]
name = "agent-security-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "typer>=0.16.0",
]

[project.optional-dependencies]
pydantic_ai = [
    "mcp==1.12.4",
    "pydantic-ai[logfire]>=0.6.2",
    "openai==1.99.1",
]

proxy = [
    "mitmproxy"
]

api = [
    "fastapi",
    "uvicorn",
    "protobuf",
    "opentelemetry-proto",
]
